Refactor working session approve test
